chore: lint fixes and resource update logic
